Fix CELL and COORDINATE mixups and cell bounds
Raw CELL values were used where a COORDINATE or TARGET is expected:

- RA's chronal vortex always appeared along the map's top edge.
- RA's fallback build placement returned a CELL, so the AI rarely built
  outside its preferred zone.
- House defeat in both games never removed the home cell's flag.
- TD's edge reinforcements headed for coordinate 0 instead of onto the map.
- The editor's Home key in both games jumped to the top edge.
- RA's dog unlimbo fallback converted the wrong way; that branch never runs,
  as documented.

Path overlap words are 32-bit, matching how they are indexed, with shared
helpers; TD's Overlap_Bit no longer shifts by -1. Adjacent_Cell checks the
neighbour index before forming the pointer. New tests cover the overlap
bits and TD's neighbour lookup.

+++ src/ra/defines.h
@@ -2827,7 +2827,7 @@ struct PathType {
   int Cost;                // Accumulated terrain cost.
   int Length;              // Command string length.
   FacingType* Command;     // Pointer to command string.
-  unsigned long* Overlap;  // Pointer to overlap list
+  uint32_t* Overlap;       // Overlap bitmap, see ra/path_overlap.h.
   CELL LastOverlap;        // stores position of last overlap
   CELL LastFixup;          // stores position of last overlap
 };
